requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.18.3
SQLAlchemy==2.0.44
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic==2.10.3
email-validator==2.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
